pyproject.toml
[project]
name = "mercati-energetici"
version = "0.1.0"
description = "Async python client for the Italian energy markets API."
authors = [{name = "Davide Marcato"}]
maintainers = [{name = "Sam Reghenzi"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yarl>=1.8.2",
    "aiohttp>=3.8.4",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "pytest-asyncio>=0.21.0",
    "aioresponses>=0.7.6",
    "mkdocs-material>=9.1.4",
    "mkdocstrings[python]>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mercati_energetici"]
